chore(#3058): move emitResizableAbExports to dataview-native.ts + refresh LOC baseline (#3102) ✓
The quality gate's LOC-regrowth ratchet flagged the +162 index.ts growth —
the exports emitter belongs in the AB/DataView subsystem module, not the
barrel. Moved verbatim (compile output byte-identical, spot-checked sha256
both lanes); index.ts keeps only the two finalize call sites (+9). The
remaining intended growth (runtime.ts arms +174 — they need the module-
private caches; dataview-native.ts +155 — the subsystem home; new-super.ts
+6 comment) is banked in the refreshed baseline, visible here for review.
