+++ scripts/loc-budget-baseline.json
@@ -1,7 +1,7 @@
 {
   "generated": "2026-07-09",
   "threshold": 1500,
-  "totalCeiling": 386034,
+  "totalCeiling": 386565,
   "files": {
     "src/codegen-linear/index.ts": 5506,
     "src/codegen-linear/runtime.ts": 3638,
@@ -15,7 +15,7 @@
     "src/codegen/class-bodies.ts": 3032,
     "src/codegen/closures.ts": 5000,
     "src/codegen/context/types.ts": 2432,
-    "src/codegen/dataview-native.ts": 3045,
+    "src/codegen/dataview-native.ts": 3200,
     "src/codegen/declarations.ts": 5389,
     "src/codegen/destructuring-params.ts": 2119,
     "src/codegen/expressions.ts": 1567,
@@ -23,10 +23,10 @@
     "src/codegen/expressions/builtins.ts": 3709,
     "src/codegen/expressions/calls.ts": 17396,
     "src/codegen/expressions/identifiers.ts": 1925,
-    "src/codegen/expressions/new-super.ts": 5597,
+    "src/codegen/expressions/new-super.ts": 5603,
     "src/codegen/expressions/unary-updates.ts": 2087,
     "src/codegen/generators-native.ts": 4687,
-    "src/codegen/index.ts": 16583,
+    "src/codegen/index.ts": 16592,
     "src/codegen/json-codec-native.ts": 2859,
     "src/codegen/literals.ts": 4228,
     "src/codegen/map-runtime.ts": 2104,
@@ -57,6 +57,6 @@
     "src/ir/lower.ts": 3532,
     "src/ir/nodes.ts": 2926,
     "src/ir/select.ts": 3223,
-    "src/runtime.ts": 15191
+    "src/runtime.ts": 15365
   }
 }
